refactor: move current visualizations to a basic collection
