+++ frontend/src/metabase/visualizations/register.js
@@ -3,24 +3,24 @@ import {
   setDefaultVisualization,
 } from "metabase/visualizations";
 
-import Scalar from "./visualizations/Scalar";
-import SmartScalar from "./visualizations/SmartScalar";
-import Progress from "./visualizations/Progress";
-import Table from "./visualizations/Table";
-import Text from "./visualizations/Text";
-import LineChart from "./visualizations/LineChart";
-import BarChart from "./visualizations/BarChart";
-import WaterfallChart from "./visualizations/WaterfallChart";
-import RowChart from "./visualizations/RowChart";
-import PieChart from "./visualizations/PieChart";
-import AreaChart from "./visualizations/AreaChart";
-import ComboChart from "./visualizations/ComboChart";
-import MapViz from "./visualizations/Map";
-import ScatterPlot from "./visualizations/ScatterPlot";
-import Funnel from "./visualizations/Funnel";
-import Gauge from "./visualizations/Gauge";
-import ObjectDetail from "./visualizations/ObjectDetail";
-import PivotTable from "./visualizations/PivotTable";
+import Scalar from "./visualizations/basic/Scalar";
+import SmartScalar from "./visualizations/basic/SmartScalar";
+import Progress from "./visualizations/basic/Progress";
+import Table from "./visualizations/basic/Table";
+import Text from "./visualizations/basic/Text";
+import LineChart from "./visualizations/basic/LineChart";
+import BarChart from "./visualizations/basic/BarChart";
+import WaterfallChart from "./visualizations/basic/WaterfallChart";
+import RowChart from "./visualizations/basic/RowChart";
+import PieChart from "./visualizations/basic/PieChart";
+import AreaChart from "./visualizations/basic/AreaChart";
+import ComboChart from "./visualizations/basic/ComboChart";
+import MapViz from "./visualizations/basic/Map";
+import ScatterPlot from "./visualizations/basic/ScatterPlot";
+import Funnel from "./visualizations/basic/Funnel";
+import Gauge from "./visualizations/basic/Gauge";
+import ObjectDetail from "./visualizations/basic/ObjectDetail";
+import PivotTable from "./visualizations/basic/PivotTable";
 
 export default function() {
   registerVisualization(Scalar);
@@ -28,6 +28,7 @@ export default function() {
   registerVisualization(Progress);
   registerVisualization(Gauge);
   registerVisualization(Table);
+  3;
   registerVisualization(Text);
   registerVisualization(LineChart);
   registerVisualization(AreaChart);
