+++ frontend/src/metabase/visualizations/visualizations/basic/Gauge.jsx
