+++ frontend/src/metabase/visualizations/visualizations/basic/LineChart.jsx
@@ -1,6 +1,6 @@
 import { t } from "ttag";
-import LineAreaBarChart from "../components/LineAreaBarChart.jsx";
-import { lineRenderer } from "../lib/LineAreaBarRenderer";
+import LineAreaBarChart from "../../components/LineAreaBarChart.jsx";
+import { lineRenderer } from "../../lib/LineAreaBarRenderer";
 
 import {
   GRAPH_DATA_SETTINGS,
@@ -9,7 +9,7 @@ import {
   GRAPH_COLORS_SETTINGS,
   GRAPH_AXIS_SETTINGS,
   GRAPH_DISPLAY_VALUES_SETTINGS,
-} from "../lib/settings/graph";
+} from "../../lib/settings/graph";
 
 export default class LineChart extends LineAreaBarChart {
   static uiName = t`Line`;
