+++ frontend/src/metabase/visualizations/visualizations/basic/ObjectDetail.jsx
