+++ frontend/src/metabase/visualizations/visualizations/basic/Progress.jsx
