+++ frontend/src/metabase/visualizations/visualizations/basic/Scalar.jsx
