+++ frontend/src/metabase/visualizations/visualizations/basic/ScatterPlot.jsx
@@ -1,6 +1,6 @@
 import { t } from "ttag";
-import LineAreaBarChart from "../components/LineAreaBarChart.jsx";
-import { scatterRenderer } from "../lib/LineAreaBarRenderer";
+import LineAreaBarChart from "../../components/LineAreaBarChart.jsx";
+import { scatterRenderer } from "../../lib/LineAreaBarRenderer";
 
 import {
   GRAPH_DATA_SETTINGS,
@@ -8,7 +8,7 @@ import {
   GRAPH_GOAL_SETTINGS,
   GRAPH_COLORS_SETTINGS,
   GRAPH_AXIS_SETTINGS,
-} from "../lib/settings/graph";
+} from "../../lib/settings/graph";
 
 export default class ScatterPlot extends LineAreaBarChart {
   static uiName = t`Scatter`;
