+++ frontend/src/metabase/visualizations/visualizations/basic/SmartScalar.jsx
