+++ frontend/src/metabase/visualizations/visualizations/basic/SmartScalar.styled.jsx
