+++ frontend/src/metabase/visualizations/visualizations/basic/Table.jsx
@@ -1,8 +1,8 @@
 /* eslint-disable react/prop-types */
 import React, { Component } from "react";
 
-import TableInteractive from "../components/TableInteractive.jsx";
-import TableSimple from "../components/TableSimple";
+import TableInteractive from "../../components/TableInteractive.jsx";
+import TableSimple from "../../components/TableSimple";
 import { t } from "ttag";
 import * as DataGrid from "metabase/lib/data_grid";
 import { findColumnIndexForColumnSetting } from "metabase/lib/dataset";
