+++ frontend/src/metabase/visualizations/visualizations/basic/WaterfallChart.jsx
@@ -1,13 +1,13 @@
 import { t } from "ttag";
-import LineAreaBarChart from "../components/LineAreaBarChart";
-import { waterfallRenderer } from "../lib/LineAreaBarRenderer";
+import LineAreaBarChart from "../../components/LineAreaBarChart";
+import { waterfallRenderer } from "../../lib/LineAreaBarRenderer";
 import { assocIn } from "icepick";
 
 import {
   GRAPH_DATA_SETTINGS,
   GRAPH_AXIS_SETTINGS,
   GRAPH_DISPLAY_VALUES_SETTINGS,
-} from "../lib/settings/graph";
+} from "../../lib/settings/graph";
 
 import { color } from "metabase/lib/colors";
